+++ webapi/src/main/scala/org/knora/webapi/slice/lists/api/ListsApiModule.scala
@@ -4,15 +4,14 @@
  */
 
 package org.knora.webapi.slice.lists.api
+
 import zio.URLayer
 import zio.ZLayer
 
 import org.knora.webapi.config.AppConfig
 import org.knora.webapi.responders.admin.ListsResponder
 import org.knora.webapi.slice.common.api.BaseEndpoints
-import org.knora.webapi.slice.common.api.HandlerMapper
 import org.knora.webapi.slice.common.api.KnoraResponseRenderer
-import org.knora.webapi.slice.common.api.TapirToPekkoInterpreter
 import org.knora.webapi.slice.lists.api.service.ListsV2RestService
 
 object ListsApiModule { self =>
@@ -20,18 +19,15 @@ object ListsApiModule { self =>
     // format: off
     AppConfig &
     BaseEndpoints &
-    HandlerMapper &
     KnoraResponseRenderer &
-    ListsResponder &
-    TapirToPekkoInterpreter
+    ListsResponder
     // format: on
 
-  type Provided = ListsApiV2Routes & ListsEndpointsV2
+  type Provided = ListsV2ServerEndpoints & ListsEndpointsV2
   val layer: URLayer[self.Dependencies, self.Provided] =
     ZLayer.makeSome[self.Dependencies, self.Provided](
-      ListsApiV2Routes.layer,
+      ListsV2ServerEndpoints.layer,
       ListsEndpointsV2.layer,
-      ListsEndpointsV2Handler.layer,
       ListsV2RestService.layer,
     )
 }