+++ webapi/src/main/scala/org/knora/webapi/slice/shacl/api/ShaclApiService.scala
@@ -7,46 +7,30 @@ package org.knora.webapi.slice.shacl.api
 
 import org.apache.jena.riot.RDFDataMgr
 import org.apache.jena.riot.RDFFormat
-import org.apache.pekko.stream.scaladsl.Source
-import org.apache.pekko.stream.scaladsl.StreamConverters
-import org.apache.pekko.util.ByteString
 import zio.*
+import zio.stream.*
 
 import java.io.FileInputStream
-import java.io.OutputStream
-import java.io.PipedInputStream
-import java.io.PipedOutputStream
 
 import org.knora.webapi.slice.shacl.domain.ShaclValidator
 import org.knora.webapi.slice.shacl.domain.ValidationOptions
 
-final case class ShaclApiService(validator: ShaclValidator) {
+final case class ShaclApiService(private val validator: ShaclValidator) {
 
-  def validate(formData: ValidationFormData): Task[Source[ByteString, Any]] = {
+  def validate(formData: ValidationFormData): Task[ZStream[Any, Throwable, Byte]] = {
     val options = ValidationOptions(
       formData.validateShapes.getOrElse(ValidationOptions.default.validateShapes),
       formData.reportDetails.getOrElse(ValidationOptions.default.reportDetails),
       formData.addBlankNodes.getOrElse(ValidationOptions.default.addBlankNodes),
     )
-    val (out, src) = makeOutputStreamAndSource()
     ZIO.scoped {
       for {
         dataStream  <- ZIO.fromAutoCloseable(ZIO.succeed(new FileInputStream(formData.`data.ttl`)))
         shaclStream <- ZIO.fromAutoCloseable(ZIO.succeed(new FileInputStream(formData.`shacl.ttl`)))
         report      <- validator.validate(dataStream, shaclStream, options)
-        _ <- ZIO.attemptBlockingIO {
-               try { RDFDataMgr.write(out, report.getModel, RDFFormat.TURTLE) }
-               finally { out.close() }
-             }.forkDaemon
-      } yield ()
-    }.as(src)
-  }
-
-  private def makeOutputStreamAndSource(): (OutputStream, Source[ByteString, _]) = {
-    val outputStream = new PipedOutputStream()
-    val inputStream  = new PipedInputStream(outputStream)
-    val source       = StreamConverters.fromInputStream(() => inputStream)
-    (outputStream, source)
+        out          = ZStream.fromOutputStreamWriter(RDFDataMgr.write(_, report.getModel, RDFFormat.TURTLE))
+      } yield out
+    }
   }
 }