+++ webapi/src/main/scala/org/knora/webapi/util/Logger.scala
@@ -4,6 +4,7 @@
  */
 
 package org.knora.webapi.util
+
 import zio.*
 import zio.logging.*
 import zio.logging.LogFormat.*
@@ -19,9 +20,10 @@ object Logger {
 
   private val logFilter = LogFilter.LogLevelByNameConfig(
     rootLogLevel,
-    ("org.apache.jena", LogLevel.Info),
+    ("org.apache.jena", LogLevel.Debug),
     ("io.netty", LogLevel.Info),
     ("org.ehcache", LogLevel.Info),
+    ("zio.http.*", LogLevel.Debug),
     // Uncomment the following lines to change the log level for specific loggers:
     // ("zio.logging.slf4j", LogLevel.Debug)
     // ("SLF4J-LOGGER", LogLevel.Warning)
@@ -50,8 +52,4 @@ object Logger {
   private val logger: ULayer[Unit] = if (useJsonLogger) jsonLogger else textLogger
 
   def fromEnv(): ULayer[Unit] = Runtime.removeDefaultLoggers >>> logger >+> Slf4jBridge.initialize
-
-  def json(): ULayer[Unit] = Runtime.removeDefaultLoggers >>> jsonLogger >+> Slf4jBridge.initialize
-
-  val text: ULayer[Unit] = Runtime.removeDefaultLoggers >>> textLogger >+> Slf4jBridge.initialize
 }